Fixed everything listed in the PR comments.

+++ src/modules/equipment/equipment.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common'
+import {
+  ConflictException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common'
 import { EquipmentRepository } from './equipment.repository'
 import { PaginatedEquipmentResult } from './dto/paginated-equipment.dto'
 import { InjectRepository } from '@nestjs/typeorm'
@@ -6,10 +11,7 @@ import { UpdateEquipmentDto } from './dto/update-equipment.dto'
 import { Equipment } from '../../database/entities/Equipment.entity'
 import { CreateEquipmentDto } from './dto/create-equipment.dto'
 import { PaginationRequestDto } from '../members/dto/pagination-request.dto'
-import {
-  calculateOffset,
-  errorMessages,
-} from '/Users/aleksa/Desktop/Projects/gym-backend/src/database/databaseUtil/utilFunctions'
+import { errorMessages } from '../../database/databaseUtil/utilFunctions'
 
 @Injectable()
 export class EquipmentService {
@@ -21,35 +23,47 @@ export class EquipmentService {
     paginationRequestDto: PaginationRequestDto
   ): Promise<PaginatedEquipmentResult> {
     try {
-      const { limit, page } = paginationRequestDto
-      const offset = calculateOffset(limit, page)
-
-      const [equipment, total] = await this.equipmentRepository.findAll(paginationRequestDto)
-      const totalPages = Math.ceil(total / limit)
-      return { data: equipment, limit, offset, total, totalPages }
+      return this.equipmentRepository.findAllEquipment({
+        limit: paginationRequestDto.limit,
+        page: paginationRequestDto.page,
+      })
     } catch (error) {
       throw new InternalServerErrorException(errorMessages.generateFetchingError('equipment'))
     }
   }
 
   async addEquipment(createEquipmentDto: CreateEquipmentDto) {
+    const existingEquipment = await this.equipmentRepository.findOneBy({
+      name: createEquipmentDto.name,
+    })
+    if (existingEquipment) {
+      throw new ConflictException('Equipment with that name already exists.')
+    }
     try {
-      const newEquipment = this.equipmentRepository.create(createEquipmentDto)
-      return await this.equipmentRepository.save(newEquipment)
+      return this.equipmentRepository.createAndSave(createEquipmentDto)
     } catch (error) {
       throw new InternalServerErrorException('Error adding equipment')
     }
   }
 
   async findByIdOrThrow(id: string): Promise<Equipment> {
-    return await this.equipmentRepository.findByIdOrThrow(id)
+    const equipment = await this.equipmentRepository.findById(id)
+    if (!equipment) throw new NotFoundException(errorMessages.generateEntityNotFound('Equipment'))
+    return equipment
   }
 
   async updateEquipment(id: string, updateEquipmentDto: UpdateEquipmentDto): Promise<Equipment> {
-    return await this.equipmentRepository.updateEquipment(id, updateEquipmentDto)
+    try {
+      return await this.equipmentRepository.updateEquipment(id, updateEquipmentDto)
+    } catch (error) {
+      throw new InternalServerErrorException(errorMessages.generateUpdateFailed('equipment'))
+    }
   }
 
   async deleteEquipment(id: string): Promise<void> {
+    if (!(await this.findByIdOrThrow(id)))
+      throw new NotFoundException(errorMessages.generateEntityNotFound('Equipment'))
+
     return await this.equipmentRepository.deleteEquipment(id)
   }
 }
